Look up owner and group thread-safely
getpwuid and getgrgid return a pointer into storage shared by the whole
process. The scan calls them for every file, from a worker, so two
lookups running at once can read each other's answer - and the archive
then records a file as belonging to somebody else. Nothing would look
wrong until a restore put the wrong owner back.

The _r forms write into a buffer the caller owns. A name that will not
resolve is left empty rather than guessed at; the numeric uid is
recorded either way, and a restore maps by name only when it recognises
one.

Verified by capturing and restoring: the manifest carries the account
name, and owner and group survive the round trip.

Also widens before adding rather than after in ArchiveCipher. Adding two
ints and casting the result would overflow in int if it overflowed at
all, and the cast would then be faithfully preserving a wrong number.

+++ src/format/crypto/ArchiveCipher.cpp
@@ -180,7 +180,11 @@ Status ArchiveCipher::encrypt(std::uint32_t blockId, ByteView plain, ByteBuffer&
         return makeError(ErrorCode::Internal, "encryption could not be finalised");
     }
 
-    const auto cipherLength = static_cast<std::size_t>(produced + finalProduced);
+    // Widened before the addition, not after: adding two ints and then casting
+    // would overflow in int if it overflowed at all, and the cast would then be
+    // faithfully preserving a wrong number.
+    const auto cipherLength =
+        static_cast<std::size_t>(produced) + static_cast<std::size_t>(finalProduced);
     if (EVP_CIPHER_CTX_ctrl(ctx.get(), EVP_CTRL_GCM_GET_TAG, static_cast<int>(kTagSize),
                             out.data() + cipherLength) != 1) {
         return makeError(ErrorCode::Internal, "could not read the authentication tag");
@@ -251,7 +255,7 @@ Status ArchiveCipher::decrypt(std::uint32_t blockId, ByteView cipher, ByteBuffer
                          "the block failed authentication: the archive is damaged or was "
                          "modified after it was written");
     }
-    out.resize(static_cast<std::size_t>(produced + finalProduced));
+    out.resize(static_cast<std::size_t>(produced) + static_cast<std::size_t>(finalProduced));
     return ok();
 #else
     (void)blockId;
